dom/base/test/test_navigator_resolve_identity.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=985827
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 985827</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 985827 **/

  // Test WebIDL NavigatorProperty objects
  var x = navigator.mozContacts;
  is(typeof x, "object", "Should have a mozContacts object");
  delete navigator.mozContacts;
  var y = navigator.mozContacts;
  is(x, y, "Should have gotten the same mozContacts object again");

  // Test Javascript-navigator-property objects
  x = navigator.mozApps;
  is(typeof x, "object", "Should have a mozApps object");
  delete navigator.mozApps;
  y = navigator.mozApps;
  is(x, y, "Should have gotten the same mozApps object again");

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=985827">Mozilla Bug 985827</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
